// app/api/master-lookup/admin-users/interface/admin-users.ts
export interface AdminUsersEntity {
  is_used?: boolean;
  is_deleted?: boolean;
  is_active: boolean;
  created_at?: string;
  updated_at?: string;
  created_by?: string;
  updated_by?: string;
  host_ip?: string;
  url?: string;
  deleted_at?: string;
  deleted_by?: string;
  tenant_id?: string;
  id: number;
  name: string;
  email: string;
  password: string;
  role: number;
}